Project_5_FlinkStreaming/Producer/requirementes.txt
KafkaProducer 
json
load_dotenv
os
load_dotenv()